move model::utils to crate::utils
